issue3 RDF/xml 형태 SPDX 문서 기반 oss notice 생성: 리팩터링(상속으로 중복로직 제거, 함수명 및 변수명 수정 등)
