+++ onot/parsing/excel.py
@@ -5,11 +5,9 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import os.path
-import re
 import openpyxl
 import pandas as pd
-from onot.parsing import spdx_license
+from onot.parsing import parser
 
 # sheets
 SHEET_DOCUMENT_INFO = "Document Info"
@@ -35,14 +33,14 @@ COLUMN_FILE_COPYRIGHT_TEXT = "File Copyright Text"
 COLUMN_ARTIFACT_OF_HOMEPAGE = "Artifact of Homepage"
 
 
-class Parser():
+class Parser(parser.AbstractParser):
 
     def __init__(self, file):
         print("debug:" + "excel Parser class")
+        super().__init__()
         self.wb = openpyxl.load_workbook(file)
-        self.doc = {}
 
-    def validate_sheet(self, file):
+    def validate_file(self, file):
         # validate file contents : all necessary information should be in the file.
         # sheet & columns
         # 1. Document Info sheet
@@ -54,13 +52,19 @@ class Parser():
         # - Package Download Location
         # - License Concluded
         # - Package Copyright Text
+        # 3. Per File Info
+        # - File Name
+        # - Artifact of Homepage
+        # - License Concluded
+        # - License Info in File
+        # - File Copyright Text
         ws_names = self.wb.sheetnames
         print("debug: " + str(ws_names))
         required_sheets = [
             SHEET_DOCUMENT_INFO,
             SHEET_PACKAGE_INFO,
-            # SHEET_EXTRACTED_LICENSE_INFO,
             SHEET_PER_FILE_INFO
+            # SHEET_EXTRACTED_LICENSE_INFO,
         ]
         for sheet in required_sheets:
             if sheet not in ws_names:
@@ -96,18 +100,6 @@ class Parser():
                     raise ValueError("email info is not existed.") 
         if "creationInfo" not in self.doc:
             raise ValueError("Organization info is not existed in the " + SHEET_DOCUMENT_INFO) 
-
-    def remove_parentheses_if_enclosed_parentheses(self, expression):
-        # if the overall expression is enclosed in parentheses, remove the parentheses.
-        if str(expression).startswith("(") and str(expression).endswith(")"):
-            return expression[1:len(expression)-1]
-        return expression
-
-    def append_package(self, package_to_be_appended):
-        # Check if there are duplicate packages already appended
-        if any(package["name"] == package_to_be_appended["name"] and package["versionInfo"] == package_to_be_appended["versionInfo"] for package in self.packages):
-            return
-        self.packages.append(package_to_be_appended)
 
     def package_info(self):
         sheet = self.wb.get_sheet_by_name(SHEET_PACKAGE_INFO)
@@ -148,29 +140,12 @@ class Parser():
             package = {
                 "name": package_info[0],
                 "versionInfo": str(package_info[1]),
-                "licenseConcluded": self.remove_parentheses_if_enclosed_parentheses(package_info[2]),
-                "licenseDeclared": self.remove_parentheses_if_enclosed_parentheses(package_info[3]),
+                "licenseConcluded": self.remove_enclosed_parentheses(package_info[2]),
+                "licenseDeclared": self.remove_enclosed_parentheses(package_info[3]),
                 "copyrightText": package_info[4],
                 "downloadLocation": package_info[5].replace("\"", ""),
             }
             self.append_package(package)
-        self.doc["packages"] = self.packages
-
-    def extract_package_name_and_package_version(self, filename):
-        def remove_meaningless_word(_filename):
-            _filename = _filename.replace("-sources", "")
-            _filename = _filename.replace("-RELEASE", "")
-            _filename = _filename.replace("-SNAPSHOT", "")
-            return _filename
-
-        filename = os.path.splitext(os.path.basename(filename))[0]
-        filename = remove_meaningless_word(filename)
-        version = re.search(r"-\d", filename)
-
-        if version:
-            return filename[:version.start()], filename[version.start()+1:]
-        else:
-            return filename, ""
 
     def per_file_info(self):
         sheet = self.wb.get_sheet_by_name(SHEET_PER_FILE_INFO)
@@ -206,17 +181,12 @@ class Parser():
             package = {
                 "name": package_name,
                 "versionInfo": str(package_version),
-                "licenseConcluded": self.remove_parentheses_if_enclosed_parentheses(per_file_info[1]),
-                "licenseDeclared": self.remove_parentheses_if_enclosed_parentheses(per_file_info[2]),
+                "licenseConcluded": self.remove_enclosed_parentheses(per_file_info[1]),
+                "licenseDeclared": self.remove_enclosed_parentheses(per_file_info[2]),
                 "copyrightText": per_file_info[3],
                 "downloadLocation": str(per_file_info[4]).replace("\"", "")
             }
             self.append_package(package)
-
-    def parse_license_expression(self, license_expression):
-        license_names = re.split(r'OR|AND|WITH', str(license_expression))
-        # remove (, ), blank
-        return list(map(lambda license: license.replace("(", "").replace(")", "").strip(), license_names))
 
     def extracted_license_info(self):
         sheet = self.wb.get_sheet_by_name(SHEET_EXTRACTED_LICENSE_INFO)
@@ -250,104 +220,4 @@ class Parser():
                 "extractedText": extracted_license_info[1],
                 "licenseName": extracted_license_info[2],
             }
-            self.extracted_licenses.append(extracted_license)
-        self.doc["extracted_license"] = self.extracted_licenses
-
-    def get_details_license(self, license_name):
-        sl = spdx_license.SPDX_License()
-
-        # check whether the licenseId is existed in the spdx license list
-        details_url = sl.get_spdx_license_detailsUrl(license_name)
-        print("debug: " + str(details_url))
-        if details_url is not None:
-            # if so, get the license text from spdx repo : "https://spdx.org/licenses/[LICENSE_ID].json"
-            details_license = sl.get_spdx_license_details(details_url)
-        else:
-            # if not, get the license text from External Refs sheet
-            if "extracted_license" not in self.doc:
-                self.extracted_license_info()
-
-            # check the license is in the Extracted License Info sheet
-            existed_in_extracted_license = False
-            for index, extracted_license in enumerate(self.doc['extracted_license']):
-                if extracted_license['identifier'] == license_name:
-                    details_license = {
-                        'licenseId': extracted_license['identifier'],
-                        'name': extracted_license['licenseName'],
-                        'licenseText': extracted_license['extractedText'],
-                        'licenseTextHtml': None
-                    }
-                    existed_in_extracted_license = True
-                    break
-            if existed_in_extracted_license is False:
-                raise ValueError("This license is not in the spdx license list. then it should be in the Extracted License Info sheet: " + license_name)
-
-                # Check if there are duplicate licenses already appended
-        return details_license
-
-    def append_details_license(self, detalis_license, package_name, package_version):
-        packages = []
-        packages.append((package_name, package_version))
-        license = {
-            "name": detalis_license['name'],
-            "licenseId": detalis_license['licenseId'],
-            "packages": packages,
-            "licenseText": detalis_license['licenseText'],
-            "licenseTextHtml": detalis_license['licenseTextHtml']
-        }
-        self.licenses.append(license)
-
-    def add_package_info_if_license_exist(self, license_name, package_name, package_version):
-        # then, update data just to add package info
-        for index, license in enumerate(self.licenses):
-            if license["licenseId"] == license_name:
-                # then, update data just to add package info
-                target_license = license
-                target_package = license['packages']
-                target_package.append((package_name, package_version))
-                target_license['packages'] = target_package
-                self.licenses[index] = target_license
-                return True
-        return False
-
-    def license_info(self):
-        self.licenses = []
-
-        # get necessary license info
-        for package in self.packages:
-            package_name = package["name"]
-            package_version = str(package["versionInfo"])
-            license_expression = package["licenseConcluded"]
-            if license_expression is None:
-                license_expression = package["licenseDeclared"]
-
-            license_names = self.parse_license_expression(license_expression)
-            for license_name in license_names:
-                if self.add_package_info_if_license_exist(license_name, package_name, package_version) is False:
-                    details_license = self.get_details_license(license_name)
-                    self.append_details_license(details_license, package_name, package_version)
-
-        self.doc["licenses"] = self.licenses
-
-    def load_doc(self, file):
-
-        # Document info
-        self.document_info()
-
-        # Package info
-        self.package_info()
-
-        # Per File Info
-        self.per_file_info()
-
-        # License info
-        self.license_info()
-
-        return self.doc
-
-    def parse(self, file):
-        print("debug: " + "parse")
-
-        self.validate_sheet(file)
-        doc = self.load_doc(file)
-        return doc
+            self.doc["extracted_license"].append(extracted_license)
